Migrate to Gradle, implement project modules
