+++ bungee/src/main/java/dev/_2lstudios/cleanmotd/bungee/commands/CleanMOTDCommand.java
