+++ bungee/src/main/java/dev/_2lstudios/cleanmotd/bungee/listeners/ProxyPingListener.java
