+++ gradle.properties
@@ -1,0 +1,5 @@
+group = dev._2lstudios.cleanmotd
+version = 0.2.7
+description = Simple MOTD managing plugin.
+url = https://github.com/2lstudios-mc/CleanMOTD
+id = cleanmotd
