+++ jar/src/main/resources/config.yml
@@ -1,0 +1,56 @@
+# Using ProtocolLib will use packets instead of events.
+motd:
+  enabled: true
+
+  # Placeholders: %onlineplayers% %maxplayers%
+  motds:
+    - |- 
+        &b&lCleanMOTD &f&lDefault &b&lMOTD
+        &7There are &a%onlineplayers%&7 players online!
+    - |-
+        &aCleanMOTD Default MOTD &c[Optimized]
+        &d&lCome and enjoy our brand new games!
+
+# This is the list of message shown when hovering over the player count.
+# Only works on Bukkit (Requires ProtocolLib) and Velocity
+sample:
+  enabled: true
+
+  # Placeholders: %onlineplayers% %maxplayers%
+  samples:
+    - |-
+      &b&lCleanMOTD &f&lDefault &b&lSample
+      &7Online: &a%onlineplayers%&8/&a%maxplayers%
+    - |-
+      &aCleanMOTD Default Sample &c[Optimized]
+      &d&lNew Sample feature!
+
+# This is shown as the server version name.
+# Requires ProtocolLib on Bukkit to work.
+protocol:
+  enabled: true
+
+  # Protocol name
+  name: "FlameCord 1.7-1.16.5"
+
+maxplayers:
+  enabled: true
+
+  # Changes the max players value.
+  maxplayers: 1400
+
+  # Changes the max player value depending on your onlineplayers.
+  # Example: 12/13 players online.
+  justonemore: true
+
+# Requires ProtocolLib on Bukkit to work.
+fakeplayers:
+  enabled: false
+
+  # Sets the amount of fakeplayers.
+  amount: 3
+
+  # "STATIC": Adds the value of fakeplayers to the amount of players online.
+  # "RANDOM": Adds a random amount of players between 1 and the fakeplayers value.
+  # "DIVISION": Adds the amount of players online divided by the fakeplayers value.
+  mode: "DIVISION"
