--- resources/bungee.yml
@@ -1,5 +1,0 @@
-name: CleanMOTD
-version: 0.2.7
-main: dev._2lstudios.cleanmotd.bungee.CleanMOTD
-description: Simple MOTD managing plugin.
-author: 2LS
