--- resources/plugin.yml
@@ -1,10 +1,0 @@
-name: CleanMOTD
-version: 0.2.7
-main: dev._2lstudios.cleanmotd.bukkit.CleanMOTD
-description: Simple MOTD managing plugin.
-author: 2LS
-softdepend: [ProtocolLib]
-api-version: 1.13
-commands:
-  cleanmotd:
-    description: "Main CleanMOTD command."
