+++ settings.gradle.kts
@@ -1,0 +1,6 @@
+rootProject.name = "cleanmotd-parent"
+
+listOf("spigot", "velocity", "bungee", "jar").forEach {
+    include("cleanmotd-$it")
+    project(":cleanmotd-$it").projectDir = file(it)
+}
