+++ spigot/src/main/java/dev/_2lstudios/cleanmotd/bukkit/commands/CleanMOTDCommand.java
