+++ spigot/src/main/java/dev/_2lstudios/cleanmotd/bukkit/listeners/ServerListPingListener.java
