+++ spigot/src/main/java/dev/_2lstudios/cleanmotd/bukkit/utils/ConfigurationUtil.java
