+++ spigot/src/main/java/dev/_2lstudios/cleanmotd/bukkit/variables/Messages.java
